requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
PyMySQL==1.1.0
bcrypt==4.0.1
python-dateutil==2.8.2
reportlab==4.0.4
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.0
cryptography==41.0.7